modules/mpris: fix dependency search
Move dependency discovery to be with other module dependency searches.
Also only define the `sdbus` dependency if the module is enabled.

The `mpris` module was always being compiled because the `sdbus`
dependency always existed. Instead, check for the external
dependency's status and create the `sdbus` bridge dependency only when
necessary.